Unpack all the sub-included .h files in create_entity.h
