+++ src/st/rwrp/e_init.c
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "rwrp.h"
-#include "../st_private.h"
 
 void OVL_EXPORT(EntityRedDoor)(Entity*);
 void OVL_EXPORT(EntityBackgroundBlock)(Entity*);
